+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2
+* Added the `Data.Proxyless` module
+* Added `proxyHashed` to `Data.Proxied`
+
 ### 0.1.1
 * Enabled `Safe`